fix(c): Update copyright
Back in February, we removed our CC license until we could think about it more carefully. This made trouble for folks that wanted to put it places like Wikipedia. Fair enough.

At their prompting, this does the careful thinking we needed to do by setting the license to CC BY-NC-SA, which should do the trick.